examples

+++ software/react_examples/src/main/scala/react/examples/husky/HuskyPathfinder.scala
@@ -19,9 +19,9 @@ class HuskyPathfinder(_id: String) extends HuskyRobot(_id) with FsmController {
   var destX = -2
   var destY = -8
 
-  var history = Set[(Long,Long)]()
+  var history = Set[(Int,Int)]()
 
-  def goalDist(x: Long, y: Long): Double = {
+  def goalDist(x: Int, y: Int): Double = {
     val dx = Math.abs(x - destX)
     val dy = Math.abs(y - destY)
     Math.sqrt(dx*dx + dy*dy)
@@ -38,7 +38,7 @@ class HuskyPathfinder(_id: String) extends HuskyRobot(_id) with FsmController {
     }
   }
 
-  var moves: List[(Long,Long)] = List()
+  var moves: List[(Int,Int)] = List()
   var actions: List[Int] = List()
 
   state('main) {
